fix: models name and fields

+++ src/models/Room.py
@@ -5,4 +5,4 @@ class Room(db.Model):
     id_room = db.Column(db.String(50), primary_key=True)
     room_type = db.Column(db.String(50), nullable=False)
     price_per_night = db.Column(db.Numeric(10, 2), nullable=False)
-    max_guests = db.Column(db.Integer, nullable=False)
+    max_guests = db.Column(db.SmallInteger, nullable=False)
